refactor so that there is a single functiont that performs tile grid_sampling
I noticed that that the function that was shifting sources in padded tiles in `render_tiles` was exactly the same function as `center_tiles` in grid (that was being used by the deblender in binary encoders)! The only difference was one line, where the offsets were the negative of each other (makes sense, one was shift away from the center and one was shifting towards the center). Since this is the most complicated function in both of the modules, I decided to consolidate them both into a single function that uses grid_sample and adda  flag `center` which controls which mode to use. I also added unit tests that check that these functions (with the respective flag) are all equal to each other.