removed the `_crop_loss_at_border` flag
As far as I can tell, this flag was always 'False' in the model in the `desc_mnras` branch and it was never changed in any of the config files. So I think it was never set to 'True' anywhere. Regardless, I never think we would like to ignore the loss in the border pading, as we want to evaluate reconstruction in full galaxies not just what's in the tile.