IMPORTANT CHANGE: loss function of deblender encoder
I realized that since we added galaxies in the padding, we need to add said padding to the reconstruction of the galaxy encoder when evaluating the loss function, otherwise the deblender will try to reconstruct the flux of galaxies that might be blended with galxies in the central square (as this will decrease the loss function). By adding the padding to the recon_mean it will penalize the encoder for over predicting flux and only target the galaxies in tiles, in the central square.