REF carefully remove all uses of extremly confusing fit_source_ptile function and its derivatives
I also added a new test in `test_render_ptiles` where I make it clear that that function was not doing anything in the case of a galaxy tiles. The real point where it was getting trimmed was in the grid_sample function, but that is confusing because the function that was using grid_sample was suppose to just shift sources but it ended up trimming padded tiles of galaxies too, but was it doing it correctly? I also removed the use in the measuing galaxy ellipticyt function because I can just create a PSF with the correct size.